OMS Package Development (Resurrect code to check GNU comiler version, build both RPM and DEB on Ubuntu platforms, set up sudoer file permissions on both RPM and DEB, don't stop OMI before OMS install), Fix InstallBuilder unit tests on UBUNTU, Fix procfsreader typedef issues on Ubuntu 12, Fix Ruby cleanup issue (lex.c not in TFS)

+++ source/code/include/scxsystemlib/procfsreader.h
@@ -94,8 +94,8 @@ protected:
         while ( true )
         {
             const char keyValueSep = ':';
-            const unsigned short MAXLINECC = (sizeof(cpuInfoProperty)/sizeof(cpuInfoProperty[0]));
-            const unsigned short MAXKEYCC = (sizeof(cpuInfoKey)/sizeof(cpuInfoKey[0]));
+            const size_t MAXLINECC = (sizeof(cpuInfoProperty)/sizeof(cpuInfoProperty[0]));
+            const size_t MAXKEYCC = (sizeof(cpuInfoKey)/sizeof(cpuInfoKey[0]));
 
             ifs.getline(cpuInfoProperty, MAXLINECC);
             if ( !ifs.good() )
